adding the changes to app.js and adding home page and pages structure. Essential upload.
