+++ client/src/Pages/Home/Home.css
@@ -1,0 +1,9 @@
+#introImgDiv {
+    height: 100vh;
+    width: 100vh;
+    z-index: 2;
+    padding-top: 150px;
+    height: 100%;
+    width: 100%;
+    z-index: 2;
+}
