+++ client/src/components/Map/GoogleMapsContainer.js
@@ -6,7 +6,8 @@ import CurrentLocation from '../SimpleMaps';
 let styles = {
   mapContainer: {
     width: "50%",
-    height: "atuo"
+    height: "atuo",
+    marginBottom: "100px"
   }
 }
